requirements.txt
# Core
langchain>=0.2.10
langchain-community>=0.2.10
langchain-openai>=0.1.22

# Vector DB
chromadb>=0.5.5

# Tokenizer
tiktoken>=0.7.0

# PDF support
pypdf>=4.2.0

# .env support
python-dotenv>=1.0.1
